Move Interleaving into package organization
